feat: Complete DataStructure#build_structure and bug fix in all possible destination method
Finished the build structure method by changing to a recursive approach and setting the @root_cell to root_cell parameter then setting a @structure to the build_structure result. Before I was trying to set the build_structure result to @root_cell. This is redundant because both @structure and @root_cell point to the same thing (I think). But I needed to do it this way because the build structure logic needs a static pointer to the initial root cell to recalculate the level order array with each recursion. It could probably be improved but it works at this point. Also fixed a minor logical error in the all possible destinations method.