Create knight moves method and update constructor
The knight moves method is now complete and with it, the project. The method takes in a target cell and then within an each loop recursively searches the structure until the target is found, at which point the loop breaks. A path array is maintained throughout the program, adding a position everytime the search descends a level and removing one every time the search goes back a level. In the end each element in the path array is printing, representing the path the knight took to reach its target.

A quick optimization was to change the constructor method to instead of assigning the build_structure return value to the structure instance variable, just remove the structure instance variable and run the build_structure method as the final part of the constructor.